dcbf_repro/requirements.txt
numpy>=1.24
scipy>=1.11
torch>=2.2
PyYAML>=6.0
tqdm>=4.66
matplotlib>=3.8
pandas>=2.2
tensorboard>=2.16
rich>=13.7
